Moved to flattening module. Renamed emittance to luminance. Implemented caching in BlockState. Reindented files.
