+++ station-blocks-v0/src/main/resources/station-blocks-v0.mixins.json
@@ -1,28 +1,23 @@
 {
-	"required": true,
-	"minVersion": "0.8",
-	"package": "net.modificationstation.stationapi.mixin.block",
-	"compatibilityLevel": "JAVA_17",
-	"mixins": [
-		"AchievementGUIDAccessor",
-		"BlockBaseAccessor",
-		"FireAccessor",
-		"ItemBaseAccessor",
-		"LevelMixin",
-		"MixinBlock",
-		"MixinBlockBase",
-		"MixinClass417",
-		"MixinFire",
-		"MixinSecondaryBlock",
-		"MixinStats",
-		"StatAccessor",
-		"StatEntityAccessor",
-		"StatsAccessor"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	},
-	"client": [
-		"MixinBlockRenderer"
-	]
+  "required": true,
+  "minVersion": "0.8",
+  "package": "net.modificationstation.stationapi.mixin.block",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "AchievementGUIDAccessor",
+    "BlockBaseAccessor",
+    "FireAccessor",
+    "ItemBaseAccessor",
+    "MixinBlock",
+    "MixinBlockBase",
+    "MixinFire",
+    "MixinSecondaryBlock",
+    "MixinStats",
+    "StatAccessor",
+    "StatEntityAccessor",
+    "StatsAccessor"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
 }
