+++ station-flattening-v0/src/main/java/net/modificationstation/stationapi/impl/block/StationFlatteningBlockInternal.java
@@ -1,0 +1,9 @@
+package net.modificationstation.stationapi.impl.block;
+
+import net.modificationstation.stationapi.api.block.BlockState;
+
+import java.util.function.ToIntFunction;
+
+public interface StationFlatteningBlockInternal {
+    ToIntFunction<BlockState> stationapi_getLuminanceProvider();
+}
