+++ station-flattening-v0/src/main/java/net/modificationstation/stationapi/mixin/flattening/MixinClass417.java
@@ -1,13 +1,11 @@
 package net.modificationstation.stationapi.mixin.flattening;
 
+import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.class_417;
 import net.minecraft.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
+import org.spongepowered.asm.mixin.injection.*;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(class_417.class)
@@ -21,7 +19,6 @@ public class MixinClass417 {
         minBlock = (short) level.getBottomY();
     }
 
-    @SuppressWarnings("MixinAnnotationTarget")
     @ModifyConstant(method = "method_1402(Lnet/minecraft/level/Level;)V", constant = @Constant(expandZeroConditions = Constant.Condition.GREATER_THAN_OR_EQUAL_TO_ZERO, ordinal = 0))
     private int changeMinHeight(int value) {
         return minBlock;
@@ -40,5 +37,17 @@ public class MixinClass417 {
     @ModifyConstant(method = "method_1402(Lnet/minecraft/level/Level;)V", constant = @Constant(intValue = 127))
     private int changeMaxHeightFallback(int value) {
         return maxBlock - 1;
+    }
+
+    @ModifyVariable(
+            method = "method_1402(Lnet/minecraft/level/Level;)V",
+            at = @At(
+                    value = "STORE",
+                    ordinal = 2
+            ),
+            index = 20
+    )
+    private int getStateLuminance(int original, @Local Level level, @Local(index = 10) int x, @Local(index = 15) int y, @Local(index = 11) int z) {
+        return level.getBlockState(x, y, z).getLuminance();
     }
 }
