+++ station-flattening-v0/src/main/java/net/modificationstation/stationapi/mixin/flattening/client/MixinBlockRenderer.java
@@ -1,4 +1,4 @@
-package net.modificationstation.stationapi.mixin.block;
+package net.modificationstation.stationapi.mixin.flattening.client;
 
 import net.minecraft.block.BlockBase;
 import net.minecraft.client.render.block.BlockRenderer;
@@ -14,14 +14,14 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(value = BlockRenderer.class, priority = 500)
 public class MixinBlockRenderer {
-	@Shadow private BlockView blockView;
-	
-	@Inject(method = "render", at = @At("HEAD"))
-	private void captureLightEmission(BlockBase block, int x, int y, int z, CallbackInfoReturnable<Boolean> info) {
-		if (blockView instanceof BlockStateView stateView) {
-			BlockState state = stateView.getBlockState(x, y, z);
-			BlockBrightness.light = state.getBlock().getEmittance(state);
-		}
-		else BlockBrightness.light = BlockBase.EMITTANCE[block.id];
-	}
+    @Shadow private BlockView blockView;
+
+    @Inject(method = "render", at = @At("HEAD"))
+    private void captureLightEmission(BlockBase block, int x, int y, int z, CallbackInfoReturnable<Boolean> info) {
+        if (blockView instanceof BlockStateView stateView) {
+            BlockState state = stateView.getBlockState(x, y, z);
+            BlockBrightness.light = state.getLuminance();
+        }
+        else BlockBrightness.light = BlockBase.EMITTANCE[block.id];
+    }
 }
